Inital commit
The script crawls a given path and looks for Courier IMAP keyword
database files. It parses them and checks if they have a kind of 'Junk'
or 'Nonjunk' IMAP keyword or tag. It gets the pathes to these emails and
should then feed sa-learn to teach spamassassin. This feature is lacking
at the moment.

Also it should be able to read a config file to set global variables or
read the commandline instead.