Made the crawling more efficient. Implemented help.
Now it doesn't crawl the whole directory for every single mail anymore. That was
insane. It's now possible to give a maildir_scheme in the config file where a
maildir name can be given. This is usefull with more than one account.

Also the search for the filename is a little bit more efficient. Mailpathes
already found are kicked out of the search.